refactor(types): move some definition in separate files
