+++ package.json
@@ -35,6 +35,7 @@
   },
   "files": [
     "lib",
+    "types",
     "index.js",
     "index.d.ts"
   ],
