+++ types/cpuMetric.d.ts
@@ -1,0 +1,10 @@
+export interface CPUMetric {
+  /**
+    * Usage percentage
+    */
+  usage: number,
+  /**
+    * Raw value returned by `process.cpuUsage()`
+    */
+  raw: NodeJS.CpuUsage
+}
