+++ types/eventLoopUtilizatioMetric.d.ts
@@ -1,0 +1,11 @@
+export interface EventLoopUtilizationMetric {
+  /**
+    * Raw metric value
+    */
+   raw: {
+      idle: number,
+      active: number,
+      utilization: number
+   }
+}
+
